backend/requirements.txt
# ── Server ────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0

# ── ML / Pipeline ────────────────────────────────────────
numpy>=1.24.0
torch>=2.1.0
transformers>=4.35.0
sentencepiece>=0.1.99
faster-whisper>=1.0.0

# ── TTS ──────────────────────────────────────────────────
edge-tts>=6.1.0
pydub>=0.25.1

# ── Testing ──────────────────────────────────────────────
pytest>=7.0.0
pytest-asyncio>=0.21.0
